Correct spelling for ParkingAuthority.sol
I corrected the spelling in lines 4 from "ParkingDOA" to  "ParkingDAO". and in line 6 from "DOA" to "DAO". I suspect this was a hidden challenge like in the movie "The Last Starfighter" to see who spotted the typo, in order to recruit for an intergalactic fleet of Starfighters. I am ready!